requirements.txt
torch>=2.2
pytorch-lightning>=2.2
transformers>=4.42
datasets>=2.19
torchmetrics>=1.4
hydra-core>=1.3
omegaconf>=2.3
wandb>=0.17
onnx>=1.16
onnxruntime>=1.18
fastapi>=0.111
uvicorn>=0.30
pydantic>=2.7
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
tqdm>=4.66
flake8>=7.1
